face-verification-app/backend/requirements.txt
fastapi
uvicorn
facenet-pytorch
torch
Pillow 